+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0-edge.4] - 2026-08-18
+
 ### Fixed
 
 - Setting up a new VPS could show an install command for the `stable` channel, which has